Postgres support (#44)
I was not able to use the [postgres](https://lib.rs/crates/postgres)
crate because it starts a new tokio runtime and then does a `block_on`.
Since this project already has its own runtime, this results in a panic.
As a result, the `Database` trait had to become async.

There are no tests. Let me know if you want me to write tests against
the DB running in compose, or how you want to approach this. (Maybe
write only E2E tests instead? Maybe no tests since it's a hackathon
project?)

As per @vkgnosis instructions, the `postgres` code is copy-pasted from
the `sqlite` module and modified to work. I'm not very happy with the
maintainability of the resulting code (which is neither here nor there,
since I'm likely not the person who has to maintain it anyway). This all
seems to be throwaway code (since this project will eventually turn into
a library... maybe...), so I don't feel that bad about it.

Tested by running manually and inspecting tables through adminer from
the compose file.